+++ src/core/views/navigation/Sidebar.js
@@ -205,7 +205,6 @@ class Sidebar extends View {
         this.groupSelectorDialog = new ModalView({
             body: searchView,
             size: 'md',
-            eyebrow: "Select Group",
             header: null,
             noBodyPadding: true,
             scrollable: false,